requirement.txt
trimesh==4.2.3
accelerate==0.28.0
mesh2sdf==1.1.0
einops==0.7.0
einx==0.1.3
omegaconf==2.3.0
opencv-python==4.9.0.80
transformers==4.43.0
numpy==1.26.4
huggingface_hub
matplotlib
gradio
spaces
torch-geometric
optimum
plyfile
timm
scikit-learn
vector-quantize-pytorch
rembg
easydict
onnxruntime
utils3d
kaolin
torchtyping
point-cloud-utils
pytorch-custom-utils
beartype